refactor(files): extract requests into own folder
