+++ src/useCases/transport/files/requests/CreateTokenForFileRequest.ts
@@ -1,0 +1,5 @@
+export interface CreateTokenForFileRequest {
+    fileId: string;
+    expiresAt?: string;
+    ephemeral?: boolean;
+}
