+++ src/useCases/transport/files/requests/CreateTokenQrCodeForFileRequest.ts
@@ -1,0 +1,4 @@
+export interface CreateTokenQrCodeForFileRequest {
+    fileId: string;
+    expiresAt?: string;
+}
