+++ src/useCases/transport/files/requests/DownloadFileRequest.ts
@@ -1,0 +1,3 @@
+export interface DownloadFileRequest {
+    id: string;
+}
