+++ src/useCases/transport/files/requests/GetFileRequest.ts
@@ -1,0 +1,3 @@
+export interface GetFileRequest {
+    id: string;
+}
