+++ src/useCases/transport/files/requests/GetFilesRequest.ts
@@ -1,0 +1,6 @@
+import { OwnerRestriction } from "../../../common";
+
+export interface GetFilesRequest {
+    query?: any;
+    ownerRestriction?: OwnerRestriction;
+}
