+++ src/useCases/transport/files/requests/LoadPeerFileRequest.ts
@@ -1,0 +1,5 @@
+export interface LoadPeerFileRequest {
+    id?: string;
+    secretKey?: string;
+    reference?: string;
+}
