+++ src/useCases/transport/files/requests/UploadOwnFileRequest.ts
@@ -1,0 +1,8 @@
+export interface UploadOwnFileRequest {
+    content: Uint8Array;
+    filename: string;
+    mimetype: string;
+    expiresAt: string;
+    title: string;
+    description?: string;
+}
